Remove references to SDK since this really isn't an SDK
The SDK terminology was always a little confusing, since you'd think
that an SDK would have more to it than just a bunch of build scripts and
a few special apps. That was only the intention for the very beginning
of the project, but now it's just a historical remnant. This removes
some references in the hopes of avoiding some confusion.