Make nerves-env.sh compatible with zsh and others
Update nerves-env.sh to call itself under Bash if used in another shell
to get the correct file location.

Tested in Bash and zsh calling directly, through a symlink, an alias and
PATH.

Equalities also fixed where needed to be compatible with zsh since bash
aliases "==" to "=" and zsh does not.