Call readelf directly for executable arch info
We previously used the host system's file command. Depending on the
host, more or less data would be printed out and we'd have to ignore
some of it. This calls the version of readelf that we include in our
toolchains so it should be much more consistent across platforms. It
also verifies less information: only the platform's architecture and any
flags (hard float vs software float/ABI) are checked. This removes the
Linux kernel header check. While the kernel header check is nice, it
currently triggers false errors for Go and Rust-generated binaries.

Fixes #138