fwup: bump to v1.2.1
This version adds some bulletproofing and additional regression tests to
meta-uuid. It also adds support for ${FWUP_META_UUID} so that the UUID
can be referenced in fwup.conf scripts. This is needed to support saving
the UUID to the Nerves.Runtime.KV metadata so that it can be accessed by
firmware update libraries.