erlinit: bump to v1.4.4
This pulls in the signal mask fix. This fixes an issue where nbtty was
waiting for SIGCHLD, but would not receive it since it inherited
erlinit's signal mask. The signal mask is restored to the default before
launching applications now so that applications don't need to force
their signal masks.