fwup: bump to v1.2.6
The difference between 1.2.5 and 1.2.6 shouldn't affect Nerves. The
changes were to fix a compiler error that doesn't affect Nerves, to fix
a bug on OSX, and improve a couple issues in shell scripts.

This patch isn't included in the existing fwup version patch since that
one has been accepted to Buildroot and will disappear on the next BR
update.