Fix scrubber missing OSX executables
It turns out that readelf fails on OSX executables so there was a
regression on being able to detect them in releases. This changes the
logic so that readelf is tried first. It gives the most reliable
information. If readelf fails, then there's likely something wrong if
the file isn't a shell script. This uses `file` to figure that out and
guesses if x86_64 is in the output or "shared library" that something is
wrong. Everything else is assumed to be ok. Note that x86_64 is checked
after the readelf check so even on x86_64 devices, this check should be
ok.

Here's a sample of the error for a bad executable:

scrub-otp-release.sh: ERROR: Unexpected executable format for 'curl'

Got:
 file:Mach-O 64-bit executable x86_64

Expecting:
 readelf:ARM;0x5000200, Version5 EABI, soft-float ABI

Here's a sample of the error for a bad library:

scrub-otp-release.sh: ERROR: Unexpected executable format for '/circuits_i2c-0.1.0/priv/i2c_nif.so'

Got:
 file:Mach-O 64-bit dynamically linked shared library x86_64

Expecting:
 readelf:ARM;0x5000200, Version5 EABI, soft-float ABI

This should fix the reports of webengine_kiosk not working and it also
fixes a particularly nasty bug where a bad .so can fail a boot even when
the library that loads it isn't in the main Shoehorn list.