openssl: bump to 1.1.0h
This pulls in a pending patch in Buildroot to update the version of
OpenSSL from 1.0.2 to 1.1.0h. This fixes what appears to be issues with
Erlang using OpenSSL engines. It also enables Erlang crypto features
like additional algorithms.

OpenSSL 1.0.2 is also feeling dated now. The OpenSSL project has
indicated that they'll stop supporting it after 2019.