Fix OTP minor version detection for asdf
The asdf shims break the logic for figuring out the OTP minor version.
This change fixes it by determining the path to the OTP_VERSION file
inside of Erlang. If that fails, then fall back to just using the major
number of telling the user that the minor versions might be different.