Update scrubber error since Nerves 1.5 causes it
This attempts to help people updating to Nerves 1.5 who didn't see the
release notes. If you don't do anything to a project when you update,
the build machine's OTP libraries get included. The scrubber detects
this. The error message didn't say that this was an issue. I removed the
Nerves 1.3 error help since that feels like it has served its duty.