nerves_initramfs: new package
This pulls down a pre-built initramfs for modifying the boot of Nerves
devices. It makes it possible to revert bad builds automatically without
assistance from the primary bootloader (i.e., Raspberry Pi bootloader)
and also to mount the rootfs via the device mapper (like for encrypted
rootfs support). Note that secure key handling is not yet possible in
the pre-built package, but encrypted filesystems are mountable for test
purposes.