busybox options: clean and add dnsd;udhcpd;ifconfig
This performs the regular Busybox config maintenance that turns off
rare commands and options that get introduced on version updates.

It also enables dnsd, udhcpd, and ifconfig. This are useful to have for
vintage_net and they "add" about 50K (assuming Busybox docs are accurate).

The result of this change is that Busybox went from 272KB to 252KB, so
it's a 20KB net win.