mesa3d: add support for v3d
This makes it possible to enable the v3d driver for the Raspberry Pi 4.
Video on the RPi4 is still not working for me, but I believe this is
patch is important for others to use. Since v3d support option has to be
enabled in Buildroot for it to alter a build, this is a low risk patch.