ply: new package
This adds the ply utility for running eBPF programs.

It requires very low level knowledge to run unfortunately, so it's not
something that will be easily accessible. To get it running, I had to:

1. Make sure it was compiled with the correct Linux headers
2. Change the following kernel options:

```
+CONFIG_BPF_SYSCALL=y
-# CONFIG_PERF_EVENTS is not set
+CONFIG_KPROBES=y
+CONFIG_FUNCTION_TRACER=y
+CONFIG_FTRACE_SYSCALLS=y
```

3. Mount debugfs at `/sys/kernel/debug`

The directions on the ply site mostly work. For the Raspberry Pi, the
Linux syscall functions start with `sys` and not `SyS` since it doesn't
use syscall wrappers like x86.