rng-tools: make jitterentropy optional
This fixes warning prints and their associated boot delays on Raspberry
Pi Zero and other systems. Specifically, the following is printed on the
console:

```
JITTER rng fails with code 2
Failed to init entropy source jitter
```

See https://bugs.busybox.net/show_bug.cgi?id=12511 for discussion.  For
the official Nerves systems, having Jitterentropy being enabled or
disabled at compile time is reasonable since it's easy to tell whether
it's in use at all. For example, it just doesn't work on the Raspberry
Pi Zero. It does work on the BBB, but there's also hardware
acceleration, so it's not really necessary (assuming the hwrng is
trustworthy, but then if it's not, do you trust jitterentropy...).