Rename busybox-1.22.config -> busybox.config
This fixes a mistake that was made years ago with keeping the version
number in the file name. This Busybox configuration only is tested
against the version of Busybox supplied by the Buildroot version
referenced in create-build.sh. It's not really possible to change the
Busybox version anyway, so specifying the version number really wasn't
useful.

This adds a symlink to avoid breaking systems that reference the old
naming of the file.