erlinit: bump to v1.7.1
This fixes an issue where if you changed the home directory in
`/etc/passwd` that `erlinit` would not pick it up. If the home directory
isn't set, the behavior is as before. If it is set, though, it will
override the default. All Nerves systems currently set it the same as
the previously hardcoded value in `erlinit`, so this doesn't change
anything for most users.