Raspberry Pi Linux 5.4 support
This updates rpi-userland and linux-firmware's Broadcom WiFi driver
installation to support Linux 5.4.

In addition to upgrading to Linux 5.4, systems will need to do the
following:

1. (all) Update the Linux kernel to Raspberry Pi's
   `raspberrypi-kernel_1.20200902-1` release.
2. (all) Set the rpi-firmware tag to match the kernel:
   ```
   BR2_PACKAGE_RPI_FIRMWARE_CUSTOM_VERSION=1.20200902
   ```
3. (Pi's with WiFi) Change all references to `pi3-miniuart-bt` to
   `miniuart-bt`. This affects `fwup.conf` and `config.txt` files.
4. (Pi's with WiFi) Deselect `BR2_RPI_WIFI_FIRMWARE` from the
   `nerves_defconfig` and add the following:
   ```
   BR2_PACKAGE_LINUX_FIRMWARE=y
   BR2_PACKAGE_LINUX_FIRMWARE_BRCM_BCM43XXX=y
   ```
5. (Pi's using the official touchscreen) The touchscreen driver changed
   names. Add the following to your Linux config:
   ```
   CONFIG_TOUCHSCREEN_RASPBERRYPI_FW=m
   ```

The update from Linux 4.19 to 5.14 appears to not require many changes
to the Linux configuration. Please compare your Linux kernel
configuration to one of the Nerves official systems to see that the
differences make sense.