Prevent PKG_CONFIG_PATH from pulling in host includes
If a user has this in their environment, it looks possible to pull in
header files and libraries from the wrong place. An issue was reported
on Slack that really sounded like this was happening and since
`PKG_CONFIG_PATH` could only cause issues, clear it out everywhere just
in case.