Work around slow path searches on MacOS
If a directory under /home was in $PATH on MacOS, the path search for
`sed` was surprisingly slow. One an M1 MBP, adding `/home/missing` to
$PATH caused `mix firmware` to regularly take >40 seconds. Without that
in the $PATH, `mix firmware` almost always takes 10 or 11 seconds.

See the conversation at
https://elixirforum.com/t/mix-firmware-takes-very-long-time-to-run-how-to-debug/41201
for much, much worse experience on a different MBP. A huge thanks to
jpalley for figuring out that the slowdown was correlated with the
`/home` directory in the path.

This PR caches the absolute path to sed to avoid searching `$PATH` each
time. On an M1 MBP, this reduces the `mix firmware` time from 40 seconds
to about 11 seconds.