Build host-erlang to ensure same host/target Erlang
For the longest time, we've had a build optimization where we reuse the
host's version of Erlang to build the target version. The requirement is
that the versions are compatible or this doesn't work. This isn't too
hard to achieve, but the new Erlang/OTP 25 JIT has a new requirement of
needing the bootstrap directory in the host Erlang.

In some sense the build times are so long that adding a few more minutes
for host-erlang isn't a big deal. It is sad, but it will make the OTP 25
JIT work and it is more "correct" in that it relies less on uncontrolled
software versions on the host.