Expose the gcc args that Buildroot uses
This will make sure that they're always included in archives. The idea
is that if there's ever any compile flag confusion, we can consult the
Nerves system archive to see what Buildroot used when building.

This will also be useful with crafting the `TARGET_GCC_FLAGS` in the
system's `mix.exs`.