nerves_heart: support include_erts: true
This installs `nerves_heart` to the staging directory so that it's
possible to use `include_erts: true` when making releases. Without this
change, the staging directory has the default version of Erlang's
`heart` that doesn't pet the hardware watchdog. That means that if you
use `include_erts: true` in your release (likely indirectly via the
Nerves tooling), then the device will reboot when the hardware watchdog
times out.