erlang: no longer install to target
This patch means that users must supply ERTS in the Mix release. This
is handled by default by the Nerves tooling. Prior to this patch, the
Mix release version of ERTS was removed so that the system-provided one
in `/usr/lib/erlang` would be used. After this patch, the
system-provided one is not installed and the Mix release ERTS is no
longer deleted.

There are a few advantages to this change:

1. It is the recommended setup for non-Nerves users. It is probably what
   most people expect these days.
2. It avoids confusion with the `$ROOTDIR` environment variable. I.e.,
   does it point to `/usr/lib/erlang` or `/srv/erlang`? Both are the
   same place now.