Add ERLANG_ERTS_VSN and clean up old Erlang versions
This adds the ERLANG_ERTS_VSN environment variable so that it's easy to
install into the ERTS directory. This will be used by nerves_heart so
that the ERTS directory does not need to be automatically detected. The
automatic detection is brittle and breaks on some unclean rebuilds and
with Buildroot 2022.02.

This commit also cleans up older Erlang versions to be the latest patch
version on that series.