rpi-distro-firmware-nonfree: add symlinks
The rpi-distro-firmware-nonfree repository really is a Debian package
script. I had sync'ed the files and symlinks in the firmware directory,
but there were a couple that I missed in the Debian post install
script.

This fixes an issue where the Raspberry Pi Zero 2W's WiFi module's
firmware wouldn't load with Linux 5.15.