Don't fail build for strip errors
Since executables are already checked for compatibility with the target
processor, it's unnecessary to fail for strip errors. If strip fails,
just move on and print out a warning. This fixes issues with precompiled
Rust binaries that shows up as the following error:

```
not enough room for program headers, try linking with -N
```

Since the Rust binaries are precompiled, you can't relink, but it's
unnecessary anyway since the binaries work fine.