Don't prompt on non-interactive terminals
When nerves_system_br or its patches change, there's a warning that you
may need to build clean. It requires an acknowledgment. When building
with a non-interactive terminal, like when being called from Elixir,
it's not possible to acknowledge it. This detects the case and prints a
warning about blindly continuing since that's often the right answer for
expert users anyway. There's a pause, though, so it's easier to see in
the flood of Buildroot compilation messages.