Remove required rng-tools dependency and add notice
This makes it possible to make a Nerves system that doesn't include
`rngd`. Linux kernels have improved since the days of low entropy hangs
at early boot. This is still something that needs care from Nerves
systems maintainers so a mandatory notice is added to make sure it
doesn't accidentally slip past anyone. However, the hope is that more
and more Nerves systems won't need this and eventually even the notice
can be removed.