Refresh CI build and parallize more to fix timeout
This is needed to beat CircleCI's 60 minute build limit. The arm64
docker build is the problematic one due to emulation. This only gets it
down to 56 minutes, but that's enough.

This update also significantly bumps the container that's doing the
build. It was a super old Alpine container, and it seems like there's no
reason anymore to do something custom when the official images contain
almost everything that's needed anyway.