Simplify CI and use arm64 runner to speed up build
The build had been timing out on CircleCI. The big optimization was to
use an arm64 runner, but saving and restoring caches also had some
overhead. The Docker images now get built and pushed in a few minutes.
The tests still take 20-30 minutes. Note that ARM64 is now tested in
parallel with amd64.